Splitted userContext into: user, userFeedback and login Contexts
