+++ src/App.jsx
@@ -1,11 +1,11 @@
+import {useContext} from 'react';
 import {Navigate, Route, Routes} from 'react-router-dom';
+import {loginContext} from 'components/contexts/login/LoginContext';
 import Login from 'views/Login';
 import Admin from 'views/Admin';
-import {useContext} from 'react';
-import {userContext} from 'components/contexts/user/UserContext';
 
 const App = () => {
-  const {loggedIn} = useContext(userContext);
+  const {loggedIn} = useContext(loginContext);
   return (
     <Routes>
       <Route path='/' element={<Admin />} />
