+++ src/components/experiences/ExperienceItem.jsx
@@ -1,12 +1,13 @@
 import {useContext, useState} from 'react';
-import {userContext} from 'components/contexts/user/UserContext';
+import {loginContext} from 'components/contexts/login/LoginContext';
+import {userFeedbackContext} from 'components/contexts/user-feedback/UserFeedbackContext';
 import CloseAndEdit from 'components/close-icon/CloseAndEdit';
 import ExperienceForm from 'components/forms/experiences/ExperienceForm';
 import styles from './ExperienceItem.module.css';
 
 export default function ExperienceItem({experience, setExperiences}) {
-  const {loggedIn, makeRequest} = useContext(userContext);
-
+  const {loggedIn} = useContext(loginContext);
+  const {makeRequest} = useContext(userFeedbackContext);
   const [showForm, setShowForm] = useState(false);
 
   if (!loggedIn && showForm) setShowForm(false);
