+++ src/components/experiences/Experiences.jsx
@@ -1,13 +1,14 @@
 import {useContext, useEffect, useState} from 'react';
-import {userContext} from 'components/contexts/user/UserContext';
+import {loginContext} from 'components/contexts/login/LoginContext';
+import {userFeedbackContext} from 'components/contexts/user-feedback/UserFeedbackContext';
 import ExperienceForm from 'components/forms/experiences/ExperienceForm';
 import Button from 'components/button/Button';
 import ExperienceItem from './ExperienceItem';
 import styles from './Experiences.module.css';
 
 export default function Experiences() {
-  const {loggedIn, makeRequest, useFetch} = useContext(userContext);
-
+  const {loggedIn} = useContext(loginContext);
+  const {makeRequest, useFetch} = useContext(userFeedbackContext);
   const [loading, experiences, setExperiences] = useFetch({url: 'experiences'});
 
   const [showNewForm, setShowNewForm] = useState(false);
