+++ src/components/skills/SkillItem.jsx
@@ -1,13 +1,14 @@
 import {useContext, useState} from 'react';
-import {userContext} from 'components/contexts/user/UserContext';
+import {loginContext} from 'components/contexts/login/LoginContext';
+import {userFeedbackContext} from 'components/contexts/user-feedback/UserFeedbackContext';
 import CloseAndEdit from 'components/close-icon/CloseAndEdit';
 import SkillForm from 'components/forms/skills/SkillForm';
 import ProgressRing from './ProgressRing';
 import styles from './SkillItem.module.css';
 
 export default function SkillItem({skill, setSkills}) {
-  const {loggedIn, makeRequest} = useContext(userContext);
-
+  const {makeRequest} = useContext(userFeedbackContext);
+  const {loggedIn} = useContext(loginContext);
   const [showForm, setShowForm] = useState(false);
 
   async function deleteSkill() {
