+++ src/components/skills/Skills.jsx
@@ -1,13 +1,14 @@
 import {useContext, useEffect, useState} from 'react';
-import {userContext} from 'components/contexts/user/UserContext';
+import {loginContext} from 'components/contexts/login/LoginContext';
+import {userFeedbackContext} from 'components/contexts/user-feedback/UserFeedbackContext';
 import SkillForm from 'components/forms/skills/SkillForm';
 import Button from 'components/button/Button';
 import SkillItem from './SkillItem';
 import styles from './Skills.module.css';
 
 export default function Skills() {
-  const {loggedIn, makeRequest, useFetch} = useContext(userContext);
-
+  const {makeRequest, useFetch} = useContext(userFeedbackContext);
+  const {loggedIn} = useContext(loginContext);
   const [loading, skills, setSkills] = useFetch({url: 'skills'});
   const [showNewForm, setShowNewForm] = useState(false);
