+++ src/components/techs-and-info/TechsAndInfo.jsx
@@ -1,11 +1,14 @@
 import {useContext, useEffect, useState} from 'react';
+import {loginContext} from 'components/contexts/login/LoginContext';
 import {userContext} from 'components/contexts/user/UserContext';
 import CloseAndEdit from 'components/close-icon/CloseAndEdit';
 import Techs from './Techs';
 import styles from './TechsAndInfo.module.css';
 
 export default function TechsAndInfo() {
-  const {loggedIn, user, setUser, loadingUser, saveUser} = useContext(userContext);
+  const {user, setUser, loadingUser, saveUser} = useContext(userContext);
+  const {loggedIn} = useContext(loginContext);
+
   const [editAboutMe, setEditAboutMe] = useState(false);
   useEffect(() => {
     if (!loggedIn) setEditAboutMe(false);
